Re-organize pipeline package into component packages
