+++ pkg/pipeline/events/file_reader.go
